Fixed generic collider class
