+++ Assets/Scripts/GenericCollider.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: abff3ad55dbf8a2479cdc115b6bd8942
-timeCreated: 1488708418
+guid: 6e6e3d189939e1949b669443a586c21f
+timeCreated: 1488703795
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
